requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.5
Django==4.2.7
django-extensions==3.2.3
djangorestframework==3.14.0
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
magic-filter==1.0.11
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pyOpenSSL==23.3.0
python-telegram-bot==20.5
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.13.2
virtualenv-clone==0.5.7
Werkzeug==3.0.1
yarl==1.9.2
